refactor : 테스트 코드 재배치 (Fixtures, KitchenPosBehaviors)
