+++ src/test/java/kitchenpos/acceptance/TableAcceptanceTest.java
@@ -6,10 +6,10 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.fixture.OrderTableFixtureFactory;
+import kitchenpos.table.fixture.OrderTableFixtureFactory;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.utils.KitchenPosBehaviors;
+import kitchenpos.acceptance.util.KitchenPosBehaviors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
