+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -5,12 +5,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.fixture.OrderTableFixtureFactory;
-import kitchenpos.fixture.TableGroupFixtureFactory;
+import kitchenpos.table.fixture.OrderTableFixtureFactory;
+import kitchenpos.table.fixture.TableGroupFixtureFactory;
 import kitchenpos.table.dto.OrderTableResponse;
 import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
-import kitchenpos.utils.KitchenPosBehaviors;
+import kitchenpos.acceptance.util.KitchenPosBehaviors;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
