+++ src/test/java/kitchenpos/menu/fixture/MenuFixtureFactory.java
@@ -1,4 +1,4 @@
-package kitchenpos.fixture;
+package kitchenpos.menu.fixture;
 
 import java.math.BigDecimal;
 import java.util.List;
