+++ src/test/java/kitchenpos/menu/fixture/MenuGroupFixtureFactory.java
@@ -1,4 +1,4 @@
-package kitchenpos.fixture;
+package kitchenpos.menu.fixture;
 
 import kitchenpos.menu.domain.MenuGroup;
