+++ src/test/java/kitchenpos/order/domain/OrderLineItemFactoryTest.java
@@ -7,9 +7,9 @@ import static org.mockito.Mockito.when;
 import java.math.BigDecimal;
 import java.util.Optional;
 import kitchenpos.ServiceTest;
-import kitchenpos.fixture.MenuGroupFixtureFactory;
-import kitchenpos.fixture.MenuProductFixtureFactory;
-import kitchenpos.fixture.ProductFixtureFactory;
+import kitchenpos.menu.fixture.MenuGroupFixtureFactory;
+import kitchenpos.menu.fixture.MenuProductFixtureFactory;
+import kitchenpos.product.fixture.ProductFixtureFactory;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
