+++ src/test/java/kitchenpos/table/util/TableApplicationBehavior.java
@@ -5,8 +5,8 @@ import static java.util.stream.Collectors.toList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
-import kitchenpos.fixture.OrderTableFixtureFactory;
-import kitchenpos.fixture.TableGroupFixtureFactory;
+import kitchenpos.table.fixture.OrderTableFixtureFactory;
+import kitchenpos.table.fixture.TableGroupFixtureFactory;
 import kitchenpos.table.application.TableGroupService;
 import kitchenpos.table.application.TableService;
 import kitchenpos.table.dto.OrderTableResponse;
